+++ Sources/PickerKit/Fonts/CustomFont.swift
@@ -3,7 +3,7 @@
 //  PickerKit
 //
 //  Created by Daniel Saidi on 2022-07-11.
-//  Copyright © 2022-2025 Kankoda. All rights reserved.
+//  Copyright © 2022-2025 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
@@ -15,7 +15,11 @@ import SwiftUI
 /// be used as an optional display name. The `family` can be
 /// used for grouping when loading multiple font variants of
 /// a single font.
-public struct CustomFont: Sendable {
+///
+/// You can use the ``font(size:)`` builder to generate font
+/// values of various sizes, or use the `Font` extensions to
+/// generate SwiftUI-specific fonts.
+public struct CustomFont: Identifiable, Sendable {
 
     /// Create a custom font from a file folder.
     ///
@@ -23,35 +27,78 @@ public struct CustomFont: Sendable {
     ///   - name: The font name.
     ///   - displayName: The font display name, by default `name`.
     ///   - family: The font family name, by default the first name segment.
-    ///   - fileName: The font file name, by default `<name>.ttf`.
-    ///   - bundle: The bundle in which the file is located.
+    ///   - fileName: The font file name, by default `name`.
+    ///   - fileExtension: The font file extension, by default `ttf`.
+    ///   - bundle: The bundle in which the font file is located.
     public init(
         name: String,
         displayName: String? = nil,
         family: String? = nil,
         fileName: String? = nil,
+        fileExtension: String? = nil,
         bundle: Bundle = .main
     ) {
         self.name = name
         self.displayName = displayName ?? name
         self.family = family ?? name.defaultFamily
-        self.fileName = fileName ?? "\(name).ttf"
+        self.fileName = fileName ?? name
+        self.fileExtension = fileExtension ?? "ttf"
         self.bundle = bundle
+        registerIfNeeded()
     }
 
+    /// The unique font identifier.
+    public var id: String { name }
+
+    /// The font name.
     public let name: String
+
+    /// The font display name.
     public let displayName: String
+
+    /// The font family name.
     public let family: String
 
-    private let fileName: String
-    private let bundle: Bundle
+    /// The font file name.
+    public let fileName: String
+
+    /// The font file extension.
+    public let fileExtension: String
+
+    /// The bundle in which the font file is located
+    public let bundle: Bundle
 
     public func font(size: CGFloat) -> FontRepresentable {
-        registerIfNeeded()
-        guard let font = tryResolveFont(size: size) else {
-            fatalError("Unable to initialize font '\(name)'")
-        }
-        return font
+        if let font = tryResolveFont(size: size) { return font }
+        fatalError("Unable to initialize font '\(name)'")
+    }
+}
+
+public extension Font {
+
+    /// Returns a ``CustomFont`` with a dynamic size.
+    static func dynamic(
+      _ font: CustomFont,
+      size: CGFloat
+    ) -> Font {
+        .custom(font.name, size: size)
+    }
+
+    /// Returns a ``CustomFont`` with a fixed size.
+    static func fixed(
+      _ font: CustomFont,
+      size: CGFloat
+    ) -> Font {
+        .custom(font.name, fixedSize: size)
+    }
+
+    /// Returns a ``CustomFont`` with a style-relative size.
+    static func relative(
+        _ font: CustomFont,
+        size: CGFloat,
+        relativeTo style: Font.TextStyle
+    ) -> Font {
+        .custom(font.name, size: size, relativeTo: style)
     }
 }
 
@@ -77,7 +124,7 @@ private extension CustomFont {
     var url: CFURL? {
         bundle.url(
             forResource: fileName,
-            withExtension: nil
+            withExtension: fileExtension
         ) as CFURL?
     }
 
@@ -87,6 +134,6 @@ private extension CustomFont {
     }
 
     func tryResolveFont(size: Double) -> FontRepresentable? {
-        FontRepresentable(name: name, size: size)
+        .init(name: name, size: size)
     }
 }